fixup! feat: Implementing the toolbarto open the pages and redisgn
